+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implement service worker refresh UI to cleanly update the PWA
+
+### Fixed
+
+- Re-add service-worker.js which got lost in the migration work
+- Fixed logic to connect/disconnect Orbit.js JSON:API source queries based on online/offline status
+
 ## [0.5.2] - 2022-08-16
 
 ### Added